requirements.txt
google-genai
google-adk
requests
python-dotenv
mcp-server-calculator